+++ test_apps/weight_tracker/lib/models/dummy_test_widget_model.dart
@@ -1,0 +1,7 @@
+// This is dummy model class that is used for testing that Framy App will work
+// with models that are not annotated with FramyModel
+class DummyTestWidgetModel {
+  final String text;
+
+  DummyTestWidgetModel(this.text);
+}